% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllele2Drops.R
\name{getAllele2Drops}
\alias{getAllele2Drops}
\title{getAllele2Drops}
\usage{
getAllele2Drops(initial_df2, InputHb)
}
\arguments{
\item{initial_df2}{Data frame with parental, stutter and allele sequences}

\item{InputHb}{Minimum value of heterozygote balance}
}
\value{
Two data frames: 1) data frame with possible Allele 2 drop-outs and 2) working data frame with new column for drop-outs
}
\description{
Help function for checking Allele 2 drop-outs
}
\details{
In the first part we obtain the summary classification of the sequences into Allele 1 and Allele 2
(coded as 1 for the first and 2 for the second).
In the second part we do a second analysis only of the allele 2 sequences.
We check drop-outs of allele 2 due to removal of sequences because the
Hb values falls out of the set threshold (inputHb)
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
