% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialDataframe.R
\name{getInitialDataframe}
\alias{getInitialDataframe}
\title{getInitialDataframe}
\usage{
getInitialDataframe(df, MinimumReads, Hb_Threshold)
}
\arguments{
\item{df}{Imported data frame from lusSTR}

\item{MinimumReads}{A numerical value for minimum read counts}

\item{Hb_Threshold}{Minimum vaue of heterozygote balance}
}
\value{
Two data frames: 1) final data frame with stutter sequences and 2) data frame with noise sequences
}
\description{
Helpfunction for building working data frame
}
\details{
set minimum reads for removing noise sequences and
set Hb for definition of true alleles.
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
