% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLUS.R
\name{getLUS_lengthReps}
\alias{getLUS_lengthReps}
\title{getLUS}
\usage{
getLUS_lengthReps(df)
}
\arguments{
\item{df}{Data frame with the most frequent stutters (n-1, n-2, n+1, n+2 and n0) obtained with selectStutters()}
}
\value{
Updated data frame with variables LUS and isLUS(MotiType)
}
\description{
Helper function that adds LUS variable and isLUS (NB! isLUS is the equivalent to MotifType, see manuscript
section 2.6.3)
}
\details{
LUS variable is the number of repeats of the longest uninterrupted stretch.
isLUS/MotifType is a binary var for the type of parental uninterrupted stretch (LUS and non-LUS output)
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
