% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifReps.R
\name{getMotifReps}
\alias{getMotifReps}
\title{getMotifReps}
\usage{
getMotifReps(tmp, brksign = c("[", "]"))
}
\arguments{
\item{tmp}{A string with bracket format e.g. "\link{ATCG}6"}

\item{brksign}{Bracket signs to use}
}
\value{
Vector with 2 elements: Extracted motif and corresponding repetitions
}
\description{
Get motifs and repetitions from short bracket format
}
\details{
Help function. Obtains motif and repeats from short bracket format
}
\examples{
\dontrun{
getMotifReps("[ATCG]6")
}
}
\author{
Oyvind Bleka <oyvble.at.hotmail.com>
}
