% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifStutterInfo.R
\name{getMotifStutterInfo}
\alias{getMotifStutterInfo}
\title{getMotifStutterInfo}
\usage{
getMotifStutterInfo(
  seqBrack,
  addSeq1Coverage = TRUE,
  addSeq2Coverage = TRUE,
  addSeq1allele_type = TRUE,
  addSeq2allele_type = TRUE,
  addSeq1SequenceID = TRUE,
  addSeq2SequenceID = TRUE,
  addSeq1ATcontent = TRUE,
  addSeq2ATcontent = TRUE,
  addStutterProportion = TRUE
)
}
\arguments{
\item{seqBrack}{matrix with STRs sequence converted into bracket format, reads, allele_type,
Sequence ID and AT-content}

\item{addSeq1Coverage}{TRUE condition for adding Seq1 read counts ("coverage") to the
final stutter table}

\item{addSeq2Coverage}{TRUE condition for adding Seq2 read counts to the final
stutter table}

\item{addStutterProportion}{TRUE condition for adding the stutter proportion to the
stutter table. Proportion of read counts Seq2 compare to the total (Seq1 + Seq2)}
}
\value{
data frame with sequences Seq1 and Seq2, number of different motifs (motifDifference),
read counts (coverage) for Seq1 and Seq2, and stutter proportion.
}
\description{
Creates final table with stutter information
}
\details{
Help function. Takes as input the short bracket format sequences.
Compares one sequence with the remaining sequences by extracting and comparing
their repetitive motifs. If equal, then number of repetitions of the motifs
will be compared and variations will be considered as stutters.

NOTE: STUTTER PRODUCTS ARE ASSUMED TO NOT EXCEED THE COMPARING SEQUENCE
}
\examples{
\dontrun{
bracketformat <- c("[ATCT]12","[ATCT]10", "[ATCT]11", "[ATCT]9")
reads <- c(869, 60, 28, 23)
allele_type <- c(1,2,3,3)
Sequence_ID <- c(227760, 227761, 227762,227763)
AT_contentVar <- c(75.00, 33.33, 64.00, 77.33)
seqBrack <- rbind(reads, allele_type, Sequence_ID, AT_contentVar)
colnames(seqBrack) <- bracketformat
getMotifStutterInfo(seqBrack)}
}
\author{
Oyvind Bleka <oyvble.at.hotmail.com>
}
