% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStuttersAllele2Pos.R
\name{getStuttersAllele2Pos}
\alias{getStuttersAllele2Pos}
\title{getStuttersAllele2Pos}
\usage{
getStuttersAllele2Pos(df1, df_ext, Hb_min, Hb_max)
}
\arguments{
\item{df1}{Input data frame with pairs of parental and stutter sequences}

\item{df_ext}{Input data frame with all parental and stutter sequences}

\item{Hb_min}{Minimum Hb value}

\item{Hb_max}{Maximum Hb value}
}
\value{
Three data frames: 1) Data frames with stutter sequences that need to be removed 2) New stutter table with removed sequences
3) Input data frame with all parental and stutter sequences and new column with Allele2_type variable
}
\description{
Help function for obtaining Allele 2 sequences in n-1 position
}
\details{
Identifies Allele 2 sequences in n-1 position when Hb is < 0.3 and removes them from
stutter table
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
