% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVariables.R
\name{helpfun_AddVariables}
\alias{helpfun_AddVariables}
\title{AddVariables}
\usage{
helpfun_AddVariables(inputDF_stutterTable, in_df, ex_df1)
}
\arguments{
\item{inputDF_stutterTable}{Input data frame with pairs of parental and stutter sequences}

\item{in_df}{Data frame with sequences included in the analysis}

\item{ex_df1}{Data frame with sequences excluded in the analysis}
}
\value{
Four data frames: 1) Data frame with all excluded sequences 2) Data frame with all included sequences
3) Data frame sequences excluded before the stutter analysis (first analytical step) 4) Input data frame with pairs of parental
and stutter sequences (inlcuidng new variable)
}
\description{
Help function for adding variable "inclusion_stutterTab" to input data frame
}
\details{
Variable "inclusion_stutterTab" indicates if a sequence has been included in the stutter analysis or not
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
