% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialStutterTable.R
\name{initialStutterTable}
\alias{initialStutterTable}
\title{initialStutterTable}
\usage{
initialStutterTable(inputDF_stutterTable)
}
\arguments{
\item{inputDF_stutterTable}{Input data frame with all parental and stutter sequences}
}
\value{
Three data frames: 1) Pairs of parental and stutter 2) All sequences included in the analysis
3) Sequences excluded from the analysis
}
\description{
Help function for obtaining the initial table with the stutter sequences
}
\details{
Obtains the initial stutter table where pairs of parental and stutter sequences are
stored for further analysis. In parallel, two tables with the included and excluded sequences are returned.
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
