% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readlusSTRFile.R
\name{readlusSTRFile}
\alias{readlusSTRFile}
\title{readlusSTRFile}
\usage{
readlusSTRFile(path, lusSTRFile)
}
\arguments{
\item{path}{Add path to lusSTRFile}

\item{lusSTRFile}{name with extension of txt input file obtained from lusSTR}
}
\value{
Data frame object prepared analysis in \code{\link{getInitialDataframe}}
}
\description{
Read output file from luSTR and prepare it for later analysis
}
\details{
Import output file obtained from lusSTR
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
