% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectStutters.R
\name{selectStutters}
\alias{selectStutters}
\title{selectStutters}
\usage{
selectStutters(stutterInit_df)
}
\arguments{
\item{stutterInit_df}{Input data frame with pairs of parental and stutter sequences}
}
\value{
Four data frames: 1) Data frame with ambiguous stutters 2) Data frame with unambiguous stutters
3) Data frame with infrequent stutters 4) Data frame with frequent stutters
}
\description{
Help function for selecting unambiguous stutters
}
\details{
Obtains stutter sequences that can only be originated from one parental allele.
\subsection{A second filtering step is applied and only most common stutters plus n0}{
}

\subsection{(n-1, n-2, n+1, n+2 and n0) are further analysed. Infrequent stutters are also stored in a separate data frame.}{
}
}
\author{
Maria Martin Agudo <maagud.at.ous-hf.no>
}
